src/requirements.txt
aiohttp==3.7.4.post0
elasticsearch==7.12.0
elasticsearch-async==6.2.0
fastapi==0.63.0
pydantic==1.8.1
uvicorn==0.13.4
aioredis-cluster==1.5.2
orjson==3.3.1
fastapi-pagination==0.7.0
backoff==1.10.0